requirements.txt
colorama==0.4.6
pycryptodome==3.21.0
pycryptodomex==3.21.0
PyGetWindow==0.0.9
pywin32==303
requests==2.31.0
typing_extensions==4.12.2
winshell==0.6
